MNT: move around AtefCfgDisplay -> DesignerDisplay
