+++ atef/widgets/config.py
@@ -8,7 +8,6 @@ import json
 import logging
 import os.path
 from functools import partial
-from pathlib import Path
 from pprint import pprint
 from typing import Any, ClassVar, Dict, List, Optional, Tuple, Type, Union
 
@@ -20,7 +19,6 @@ from qtpy.QtWidgets import (QAction, QComboBox, QFileDialog, QFormLayout,
                             QMainWindow, QMessageBox, QPlainTextEdit,
                             QPushButton, QTabWidget, QToolButton, QTreeWidget,
                             QTreeWidgetItem, QVBoxLayout, QWidget)
-from qtpy.uic import loadUiType
 
 from ..check import (Comparison, Configuration, ConfigurationFile,
                      DeviceConfiguration, Equals, IdentifierAndComparison,
@@ -28,32 +26,12 @@ from ..check import (Comparison, Configuration, ConfigurationFile,
 from ..enums import Severity
 from ..qt_helpers import QDataclassBridge, QDataclassList
 from ..reduce import ReduceMethod
+from .core import DesignerDisplay
 
 logger = logging.getLogger(__name__)
 
 
-class AtefCfgDisplay:
-    """Helper class for loading the .ui files and adding logic."""
-    filename: str
-
-    def __init_subclass__(cls):
-        """Read the file when the class is created"""
-        super().__init_subclass__()
-        cls.ui_form, _ = loadUiType(
-            str(Path(__file__).parent.parent / 'ui' / cls.filename)
-        )
-
-    def __init__(self, *args, **kwargs):
-        """Apply the file to this widget when the instance is created"""
-        super().__init__(*args, **kwargs)
-        self.ui_form.setupUi(self, self)
-
-    def retranslateUi(self, *args, **kwargs):
-        """Required function for setupUi to work in __init__"""
-        self.ui_form.retranslateUi(self, *args, **kwargs)
-
-
-class Window(AtefCfgDisplay, QMainWindow):
+class Window(DesignerDisplay, QMainWindow):
     """
     Main atef config window
 
@@ -233,7 +211,7 @@ class Window(AtefCfgDisplay, QMainWindow):
         pprint(self.serialize_tree(self.get_current_tree()))
 
 
-class Tree(AtefCfgDisplay, QWidget):
+class Tree(DesignerDisplay, QWidget):
     """
     The main per-file widget as a "native" view into the file.
 
@@ -347,7 +325,7 @@ class AtefItem(QTreeWidgetItem):
         return self.widget_cached
 
 
-class Overview(AtefCfgDisplay, QWidget):
+class Overview(DesignerDisplay, QWidget):
     """
     A view of all the top-level "Configuration" objects.
 
@@ -614,7 +592,7 @@ class ConfigTextMixin:
         self.desc_edit.setFixedHeight(line_count * 13 + 12)
 
 
-class OverviewRow(ConfigTextMixin, AtefCfgDisplay, QWidget):
+class OverviewRow(ConfigTextMixin, DesignerDisplay, QWidget):
     """
     A single row in the overview widget.
 
@@ -743,7 +721,7 @@ class OverviewRow(ConfigTextMixin, AtefCfgDisplay, QWidget):
         self.overview.delete_row(self)
 
 
-class Group(ConfigTextMixin, AtefCfgDisplay, QWidget):
+class Group(ConfigTextMixin, DesignerDisplay, QWidget):
     """
     The group of checklists and devices associated with a Configuration.
 
@@ -1172,7 +1150,7 @@ class NamedDataclassList(StrList):
         self.data_list.put_to_index(index, new_bridge.data)
 
 
-class StrListElem(AtefCfgDisplay, QWidget):
+class StrListElem(DesignerDisplay, QWidget):
     """
     A single element for the StrList widget.
 
@@ -1234,7 +1212,7 @@ class FrameOnEditFilter(QObject):
         return False
 
 
-class IdAndCompWidget(ConfigTextMixin, AtefCfgDisplay, QWidget):
+class IdAndCompWidget(ConfigTextMixin, DesignerDisplay, QWidget):
     """
     A widget to manage the ids and comparisons associated with a checklist.
 
@@ -1419,7 +1397,7 @@ class IdAndCompWidget(ConfigTextMixin, AtefCfgDisplay, QWidget):
         bridge.deleteLater()
 
 
-class CompView(ConfigTextMixin, AtefCfgDisplay, QWidget):
+class CompView(ConfigTextMixin, DesignerDisplay, QWidget):
     """
     Widget to view and edit a single Comparison subclass.
